move common functions from chainbase, framewrok to common module
