+++ common/src/main/java/org/tron/common/crypto/ECKey.java
@@ -19,7 +19,7 @@ package org.tron.common.crypto;
 
 import static org.tron.common.utils.BIUtil.isLessThan;
 import static org.tron.common.utils.ByteUtil.bigIntegerToBytes;
-import static org.tron.common.utils.Commons.computeAddress;
+import static org.tron.common.utils.DecodeUtil.computeAddress;
 
 import java.io.IOException;
 import java.io.Serializable;
