--- common/src/main/java/org/tron/common/utils/Constant.java
@@ -1,26 +1,0 @@
-/*
- * java-tron is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * java-tron is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.tron.common.utils;
-
-public class Constant {
-
-  public static final long PB_DEFAULT_ENERGY_LIMIT = 0L;
-  public static final long CREATOR_DEFAULT_ENERGY_LIMIT = 1000 * 10_000L;
-
-
-  // Numbers
-  public static final int ONE_HUNDRED = 100;
-}
