+++ common/src/main/java/org/tron/common/utils/DecodeUtil.java
@@ -1,0 +1,93 @@
+package org.tron.common.utils;
+
+import static org.tron.common.utils.Hash.sha3omit12;
+import static org.tron.core.Constant.ADD_PRE_FIX_BYTE_MAINNET;
+
+import java.util.Arrays;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.ArrayUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.spongycastle.math.ec.ECPoint;
+
+@Slf4j(topic = "Commons")
+public class DecodeUtil {
+
+  public static final int ADDRESS_SIZE = 42;
+  public static final int ASSET_ISSUE_COUNT_LIMIT_MAX = 1000;
+  public static byte addressPreFixByte = ADD_PRE_FIX_BYTE_MAINNET;
+
+  public static byte[] clone(byte[] value) {
+    byte[] clone = new byte[value.length];
+    System.arraycopy(value, 0, clone, 0, value.length);
+    return clone;
+  }
+
+  private static byte[] decode58Check(String input) {
+    byte[] decodeCheck = Base58.decode(input);
+    if (decodeCheck.length <= 4) {
+      return null;
+    }
+    byte[] decodeData = new byte[decodeCheck.length - 4];
+    System.arraycopy(decodeCheck, 0, decodeData, 0, decodeData.length);
+    byte[] hash0 = Sha256Hash.hash(decodeData);
+    byte[] hash1 = Sha256Hash.hash(hash0);
+    if (hash1[0] == decodeCheck[decodeData.length] &&
+        hash1[1] == decodeCheck[decodeData.length + 1] &&
+        hash1[2] == decodeCheck[decodeData.length + 2] &&
+        hash1[3] == decodeCheck[decodeData.length + 3]) {
+      return decodeData;
+    }
+    return null;
+  }
+
+  public static boolean addressValid(byte[] address) {
+    if (ArrayUtils.isEmpty(address)) {
+      logger.warn("Warning: Address is empty !!");
+      return false;
+    }
+    if (address.length != ADDRESS_SIZE / 2) {
+      logger.warn(
+          "Warning: Address length need " + ADDRESS_SIZE + " but " + address.length
+              + " !!");
+      return false;
+    }
+
+    if (address[0] != addressPreFixByte) {
+      logger.warn("Warning: Address need prefix with " + addressPreFixByte + " but "
+          + address[0] + " !!");
+      return false;
+    }
+    //Other rule;
+    return true;
+  }
+
+  public static byte[] decodeFromBase58Check(String addressBase58) {
+    if (StringUtils.isEmpty(addressBase58)) {
+      logger.warn("Warning: Address is empty !!");
+      return null;
+    }
+    byte[] address = decode58Check(addressBase58);
+    if (address == null) {
+      return null;
+    }
+
+    if (!addressValid(address)) {
+      return null;
+    }
+
+    return address;
+  }
+
+  public static String createReadableString(byte[] bytes) {
+    return ByteArray.toHexString(bytes);
+  }
+
+  public static byte[] computeAddress(ECPoint pubPoint) {
+    return computeAddress(pubPoint.getEncoded(/* uncompressed */ false));
+  }
+
+  public static byte[] computeAddress(byte[] pubBytes) {
+    return sha3omit12(
+        Arrays.copyOfRange(pubBytes, 1, pubBytes.length));
+  }
+}
